LayoutTests/editing/pasteboard/paste-table-cells-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of BODY > HTML > #document to 5 of BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldInsertNode:#document-fragment replacingDOMRange:range from 0 of #text > DIV > BODY > HTML > #document to 9 of #text > DIV > BODY > HTML > #document givenAction:WebViewInsertActionPasted
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 0 of DIV > BODY > HTML > #document to 0 of DIV > BODY > HTML > #document toDOMRange:range from 5 of #text > TD > TR > TBODY > TABLE > DIV > BODY > HTML > #document to 5 of #text > TD > TR > TBODY > TABLE > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification

After copy:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         id="test1"
|         "<#selection-anchor>hello"
|       <td>
|         id="test2"
|         "world<#selection-focus>"
| "\n"
| <div>
|   id="target"
|   "replaceme"
| "\n\n\n"

Before paste:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         id="test1"
|         "hello"
|       <td>
|         id="test2"
|         "world"
| "\n"
| <div>
|   id="target"
|   "<#selection-anchor>replaceme<#selection-focus>"
| "\n\n\n"

After paste:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         id="test1"
|         "hello"
|       <td>
|         id="test2"
|         "world"
| "\n"
| <div>
|   id="target"
|   <table>
|     <tbody>
|       <tr>
|         <td>
|           id="test1"
|           "hello"
|         <td>
|           id="test2"
|           "world<#selection-caret>"
| "\n\n\n"
